feat(audio) : refactor out html 5 backend from player.ts
